:recycle: server type 분리
