+++ server/tsconfig.json
@@ -17,7 +17,8 @@
       "@socket/*": ["src/socket/*"],
       "@constants/*": ["src/constants/*"],
       "@api/*": ["src/api/*"],
-      "@utils/*": ["src/utils/*"]
+      "@utils/*": ["src/utils/*"],
+      "@types/*": ["src/@types/*"]
     }
   },
   "include": ["src"],
